Add CNF support and fix vector nbytes types
Replace Py_ssize_t with uint64_t in vector nbytes implementations to
ensure consistent sizing and safe casts. Add NodeType.CNF plus CNF
evaluation and optimizer handling (folding, simplification, and n-ary
flattening of OR→CNF). Extend join rewriter to recognise the LEFT ANTI
JOIN idiom and add helper utilities. Add curl-devel fallback to the
manylinux wheel build script, move TPCH query fixtures and update the
TPCH test runner path, and apply small include/import fixes.

+++ draken/vectors/bool_vector.pyx
@@ -785,8 +785,8 @@ cdef class BoolVector(Vector):
     def nbytes(self):
         """Return the approximate memory footprint of this vector in bytes."""
         cdef DrakenFixedBuffer* ptr = self.ptr
-        cdef Py_ssize_t n = ptr.length
-        cdef Py_ssize_t data_bytes, bm_bytes
+        cdef uint64_t n = ptr.length
+        cdef uint64_t data_bytes, bm_bytes
         if self._has_const:
             return 1  # single bool value
         # Bit-packed: 1 bit per element
